RemoteControl: Fix warnings

+++ plugins/RemoteControl/src/ObjectService.cpp
@@ -162,7 +162,7 @@ void ObjectService::get(const QByteArray& operation, const APIParameters &parame
 
 			// We make use of 2 extra values for linked applications. These govern sky brightness and can be used for ambient settings.
 			LandscapeMgr *lmgr=GETSTELMODULE(LandscapeMgr);
-			infoObj.insert("ambientLum", QJsonValue(lmgr->getAtmosphereAverageLuminance()));
+			infoObj.insert("ambientLum", QJsonValue(static_cast<double>(lmgr->getAtmosphereAverageLuminance())));
 			infoObj.insert("ambientInt", QJsonValue(lmgr->getCurrentLandscape()->getBrightness()));
 			response.writeJSON(QJsonDocument(infoObj));
 		}
